feat: FFT function used in worker ts
