+++ package.json
@@ -71,8 +71,10 @@
   },
   "packageManager": "yarn@1.22.22",
   "devDependencies": {
+    "@rollup/plugin-node-resolve": "^16.0.1",
     "@rollup/plugin-terser": "^0.4.4",
     "@rollup/plugin-typescript": "^12.1.1",
+    "@types/jest": "^29.5.2",
     "@typescript-eslint/eslint-plugin": "^8.18.0",
     "@typescript-eslint/parser": "^8.18.0",
     "cypress": "^13.16.1",
@@ -81,13 +83,12 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
     "glob": "^11.0.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "prettier": "^3.4.2",
     "rollup": "^4.28.1",
     "rollup-plugin-dts": "^6.1.0",
-    "typescript": "^5.7.2",
-    "jest": "^29.7.0",
     "ts-jest": "^29.1.1",
-    "@types/jest": "^29.5.2",
-    "jest-environment-jsdom": "^29.7.0"
+    "typescript": "^5.7.2"
   }
 }
